+++ tutor/commands/k8s.py
@@ -3,7 +3,7 @@ import click
 from .. import config as tutor_config
 from .. import env as tutor_env
 from .. import fmt
-from .. import interactive
+from .. import interactive as interactive_config
 from .. import opts
 from .. import scripts
 from .. import utils
@@ -16,10 +16,10 @@ def k8s():
 
 @click.command(help="Configure and run Open edX from scratch")
 @opts.root
-@click.option("-y", "--yes", "silent", is_flag=True, help="Run non-interactively")
-def quickstart(root, silent):
+@click.option("-I", "--non-interactive", is_flag=True, help="Run non-interactively")
+def quickstart(root, non_interactive):
     click.echo(fmt.title("Interactive platform configuration"))
-    config = interactive.update(root, silent=silent)
+    config = interactive_config.update(root, interactive=(not non_interactive))
     click.echo(fmt.title("Updating the current environment"))
     tutor_env.save(root, config)
     click.echo(fmt.title("Starting the platform"))